Vncserver singularity

+++ vncserver/start-template-v3.sh
@@ -122,10 +122,19 @@ if [ -z "${service_vnc_exec}" ]; then
 fi
 
 if [ -z ${service_vnc_exec} ] || ! [ -f "${service_vnc_exec}" ]; then
+    if [[ ${service_download_vncserver_container} != "true" ]]; then
+        echo "$(date) ERROR: No vncserver command found"
+        exit 1
+    fi
+    if ! which singularity > /dev/null 2>&1; then
+        echo "(date) ERROR: No vncserver or singularity command found"
+        exit 1
+    fi
     echo "$(date): vncserver is not installed. Using singularity container..."
     service_vnc_exec="singularity exec --bind /tmp/.X11-unix:/tmp/.X11-unix --bind ${HOME}:${HOME} ${service_vncserver_sif} vncserver"
     service_vnc_type="TurboVNC"
     service_desktop="echo Starting no service desktop on the host"
+    mkdir -p /tmp/.X11-unix
     rm -f ~/.vnc/xstartup.turbovnc
 cat >> ~/.vnc/xstartup.turbovnc <<HERE
 unset SESSION_MANAGER
